[#78] Move the mailbox OAuth exchange into Common so the command needs no Infrastructure
The administration command performs the authorization and the run-time adapter refreshes the grant it produced, so both
speak to one authorization server about one account and have to agree on the token response, on the sanitizing of what
that server returns, and on the endpoints each provider publishes. Reaching that code through Infrastructure would put
EF Core, Npgsql, and MailKit into an artifact that calls none of them and would make trimming it impossible, which is
the constraint the published binary is built under; duplicating it would let the two halves of one conversation drift.

Common therefore gains MailboxOAuth/ and a reference to Domain, which is what lets a failure it raises derive from
MailFathomException like every other. The authorizer's tests move with it, and the contract tests split along the same
line: what a provider preset, a PKCE challenge, and the error sanitizer decide is asserted in Common.UnitTests, and the
grant parsing and the MailKit mechanism selection stay where their subjects are.

The command itself is now wired into the mfctl command tree: mailbox authorize reaches the terminal and the network
through CliContext like every other command, so its prompts and its failures are assertable, and CliFailure carries what
it reports instead of a bare exit code.
