+++ src/Common/Common.csproj
@@ -1,13 +1,22 @@
 <Project Sdk="Microsoft.NET.Sdk">
-  <!-- Cross-cutting code that belongs to no boundary and depends on nothing but the base class library. It is
-       deliberately not a layer: nothing here knows about mail, storage, or transport, which is what lets the trimmed
-       single-file command and the service both consume it without either dragging in the other's world.
+  <!-- Cross-cutting code that belongs to no boundary and depends on nothing but the base class library and Domain. It
+       is deliberately not a layer: nothing here opens a database, reads configuration, or speaks a mail protocol, which
+       is what lets the trimmed single-file command and the service both consume it without either dragging in the
+       other's world.
+
+       Domain is the one reference, and it is not an exception to that: it is the innermost layer, it depends on nothing
+       itself, and every failure this repository publishes derives from MailFathomException. Code that cannot report a
+       failure without inventing a second hierarchy does not belong in a shared project at all.
 
        A project rather than a linked source file under src/shared/, because docs/architecture/solution-structure.md
        limits that directory to markers that carry no behavior and sends anything with executable logic here.
 
        Admission is narrow on purpose, so this does not become the drawer everything ends up in: code arrives here when
        a second boundary genuinely needs it, and until then it lives with its one consumer. -->
+  <ItemGroup>
+    <ProjectReference Include="..\Domain\Domain.csproj" />
+  </ItemGroup>
+
   <ItemGroup>
     <InternalsVisibleTo Include="MailFathom.Common.UnitTests" />
   </ItemGroup>
