+++ src/Common/MailboxOAuth/MailOAuthTokenResponse.cs
@@ -2,9 +2,10 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 // Project repository: https://github.com/Krzysztof318/MailFathom
 
+using System.Diagnostics.CodeAnalysis;
 using System.Text.Json.Serialization;
 
-namespace MailFathom.Infrastructure.Mail.OAuth;
+namespace MailFathom.Common.MailboxOAuth;
 
 /// <summary>The subset of an RFC 6749 token endpoint response MailFathom reads.</summary>
 /// <param name="AccessToken">The issued bearer token.</param>
@@ -23,7 +24,7 @@ namespace MailFathom.Infrastructure.Mail.OAuth;
 /// field nobody controls.
 /// </para>
 /// </remarks>
-internal sealed record MailOAuthTokenResponse(
+public sealed record MailOAuthTokenResponse(
     [property: JsonPropertyName("access_token")] string? AccessToken,
     [property: JsonPropertyName("expires_in")] int? ExpiresInSeconds,
     [property: JsonPropertyName("error")] string? Error,
@@ -42,7 +43,9 @@ internal sealed record MailOAuthTokenResponse(
 /// <param name="ExpiresInSeconds">How long the device code stays pollable.</param>
 /// <param name="IntervalSeconds">The minimum seconds between polls, defaulting to the RFC's 5 when absent.</param>
 /// <param name="Error">The error code from a rejected request, absent from a successful one.</param>
-internal sealed record MailOAuthDeviceAuthorizationResponse(
+[SuppressMessage("Design", "CA1054:URI-like parameters should not be strings", Justification = "The value is a JSON field read verbatim from an authorization server this process does not own. Binding it as a Uri would move a malformed value's failure inside deserialization, where it arrives as a format exception rather than as the sanitized authorization failure the caller reports.")]
+[SuppressMessage("Design", "CA1056:URI-like properties should not be strings", Justification = "The value is a JSON field read verbatim from an authorization server this process does not own. Binding it as a Uri would move a malformed value's failure inside deserialization, where it arrives as a format exception rather than as the sanitized authorization failure the caller reports.")]
+public sealed record MailOAuthDeviceAuthorizationResponse(
     [property: JsonPropertyName("device_code")] string? DeviceCode,
     [property: JsonPropertyName("user_code")] string? UserCode,
     [property: JsonPropertyName("verification_uri")] string? VerificationUri,
@@ -60,4 +63,4 @@ internal sealed record MailOAuthDeviceAuthorizationResponse(
 [JsonSerializable(typeof(MailOAuthTokenResponse))]
 [JsonSerializable(typeof(MailOAuthDeviceAuthorizationResponse))]
 [JsonSourceGenerationOptions(PropertyNameCaseInsensitive = true)]
-internal sealed partial class MailOAuthJsonContext : JsonSerializerContext;
+public sealed partial class MailOAuthJsonContext : JsonSerializerContext;
