+++ src/Common/MailboxOAuth/MailProviderPreset.cs
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 // Project repository: https://github.com/Krzysztof318/MailFathom
 
-namespace MailFathom.Infrastructure.Mail.OAuth.Authorization;
+namespace MailFathom.Common.MailboxOAuth;
 
 /// <summary>A mail provider's published OAuth endpoints and the scope its IMAP access needs.</summary>
 /// <remarks>
