+++ src/Infrastructure/Mail/OAuth/MailAccessTokenUnavailableException.cs
@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 // Project repository: https://github.com/Krzysztof318/MailFathom
 
+using MailFathom.Common.MailboxOAuth;
 using MailFathom.Domain.Failures;
 
 namespace MailFathom.Infrastructure.Mail.OAuth;
